+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Show feature name on map hover.
 - Added map symbol size override support per unit.
+- Added a button to the unit filters panel to quickly collapse or expand all filter sections.
 
 ### Changed